Separate bits of build.sh out into deploy.sh
Because we want to re-deploy the artifacts from the stage release for
the live release, we need a separate deploy.sh, with build.sh just
doing the build.  And anything that depends on REAL_WEB_URL needs to
happen in deploy.sh.

Also, remove the "prepare" dependency from some targets in
rabbitmq-umbrella/Makefile: "prepare" only needs to occur in the
context of build.sh, not deploy.sh, and so it only needs to be a
dependency of "packages", not the various other sub-targets that are
also invoked during deployment.