Don't use directories as makefile checkout targets
It's a bad idea, due to how mtimes of directories work.  Instead, use
"stamp files" to represent completed checkouts.

Don't list checkouts as dependencies of miscellaneous targets;
instead, include 'checkout' as a dependency of 'prepare'.  The
makefile isn't intended to be used for partial builds, so this apprach
is more concise, and makes more sense in my mind.

Make a start on delaring phony targets, with the .PHONY lines
occuring next to the relevant targets for maintainability.