Makefile overhaul
Introduce a uniform naming scheme for targets.

Mark targets as phony, wherever appropriate.

Replace some recursive makes with dependencies.

Update and rework the "clean" target.  autobuild/build.sh now does a
'make clean' before building.

Never do "make clean" in subdirectories while in the process of
building.

Replace some dependencies on specific files with dependencies on phony
targets.  In particular, replace the dependencies on the
rabbitmq-server source .tar.gz and .zip files with dependencies on
rabbitmq-server-srcdist, so that we only do 'make -C rabbitmq-server
srcdist' once during the build.

Topologically sort the targets.

Hopefully the end result is a much more readable Makefile, and one
which forms a good basis for further improvements in the future.