Use VERSION to refer to the global version, PACKAGE_VERSION for the package
Previously, the VERSION variable was used when invoking the main
Makefile, but the GLOBAL_VERSION variable used by packages (except for
rabbitmq-server and rabbitmq-erlang-client, which use VERSION).  This
was all a bit fragile.  It's much cleaner and robust to use VERSION
for the global version everwhere, and PACKAGE_VERSION for the version
of the package in hand.