Rebuild appropriately when VERSION or PACKAGE_VERSION changes
Also, move the generated app file from ebin/ to build/, as its
location in ebin seems a bit of a hangover from older times - the
package ebin dir is not the ebin dir we use when creating the app.
Putting it under build makes it easier to clean any versioned app
files that may be lying around.