Makefile: Fix "make up_c" with detached HEAD
When checking out a tag, a remote branch or a specific commit (ie.
something not being the HEAD of a local branch), Git operates in
detached HEAD. Before, we didn't verify this case and tried to "git pull
--ff-only" which is invalid.

Now, we check if we are in detached HEAD mode and only "git pull" if we
are not.

While here, fix a missing quote in "make checkin".

Fixes #6.