New script to check if a branch is merged before a release
This script is meant to be called with "foreachrepo". Here is an
example:
    ./foreachrepo $PWD/release-build/is-merged.sh stable \
    rabbitmq-public-umbrella#7

This example will verify is the branch "rabbitmq-public-umbrella#7" was
merged to "stable" in the umbrella and all subrepositories, if the
branch exists in those.

Therefore, it's useful if an issue required changes to the broker and
the testsuite, but not to other repositories.

Fixes #7.