build-community-plugins.py: Fix the build of all plugins
With no plugins specified on the command line, the script passed the
PLUGINS structure directly to the build() functions. Unfortunately, this
function changed in a4853f8a2764632234bca3294299eacaa53c80cc to support
both the old build system and Erlang.mk, but the "build all plugins"
code path was not changed accordingly.

This patch fixes this bug by converting the structure to a list of
tuples, as expected by build().