Go back to Erlang R16B03 to build 3.6.x releases
This is required because modules such as `rabbit_core_metrics` built
with Erlang 18.3 can't be recompiled on the fly with Erlang R16B or
17.x.

FTR, some modules need to recompile themselves on the fly to be able to
use some newer APIs such as the Time and Time Correction API or
ets:update_counter/4.

This problem was detected after the release of RabbitMQ 3.6.7. Users of
our Debian package installed with Erlang R16B03 or 17.x would see
RabbitMQ processes crash with the `{cannot_compile_forms, error}`
exception.